dto extraction done
