+++ rest-api/src/main/java/be/xplore/notifyme/controller/OrganisationManagerController.java
@@ -4,8 +4,8 @@ import be.xplore.notifyme.domain.TeamApplication;
 import be.xplore.notifyme.domain.TeamApplicationKey;
 import be.xplore.notifyme.dto.organisationapplication.ApplicationUsernameDto;
 import be.xplore.notifyme.dto.organisation.OrganisationDto;
-import be.xplore.notifyme.dto.OrganisationLimitedInfoDto;
-import be.xplore.notifyme.dto.UserApplicationResponseDto;
+import be.xplore.notifyme.dto.organisation.OrganisationLimitedInfoDto;
+import be.xplore.notifyme.dto.userapplication.UserApplicationResponseDto;
 import be.xplore.notifyme.services.ITeamApplicationService;
 import be.xplore.notifyme.services.IUserOrgApplicationService;
 import be.xplore.notifyme.services.IUserOrgService;
