+++ rest-api/src/main/java/be/xplore/notifyme/dto/organisation/OrganisationsLimitedInfoDto.java
@@ -1,4 +1,4 @@
-package be.xplore.notifyme.dto;
+package be.xplore.notifyme.dto.organisation;
 
 import be.xplore.notifyme.domain.Organisation;
 import java.util.ArrayList;
